Moved vectors to 'vector' folder
